Fix issue #154 (#155)
 Fix issue #154, num_job was a float in py2.

+++ dask_jobqueue/core.py
@@ -299,7 +299,7 @@ class JobQueueCluster(Cluster):
     def start_workers(self, n=1):
         """ Start workers and point them to our local scheduler """
         logger.debug('starting %s workers' % n)
-        num_jobs = math.ceil(n / self.worker_processes)
+        num_jobs = int(math.ceil(n / self.worker_processes))
         for _ in range(num_jobs):
             with self.job_file() as fn:
                 out = self._submit_job(fn)
